+++ agentteams_connector/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 """AgentTeams QwenPaw Workbench backend package."""
 
-__version__ = "0.5.0-beta.12.4"
+__version__ = "0.5.0-beta.12.5"